[planner] Clean up formulation of Differential IK
Resolves #9773.

There were a number of issues with the previous implementation,
discussed at some length in #9773.

One implication of the cleaner formulation is that it is possible that
the resulting optimization problem is an LP instead of a QP. In this
case, Clp is used as a solver instead of Osqp, because Osqp cannot
solve linear programs.

This change deprecates the "unconstrained degrees of freedom velocity
limit" methods of the diff IK parameters class. This limit was only
needed before because the formulation was bad. Since I'm already
changing things on users, I've taken the chance to also deprecate the
"timestep" methods in favor of "time_step", which we have agreed is
the preferred form in Drake.

This PR also changes the default arguments for the Diff IK Paremeters
class constructor; setting the default velocities to be the same as
the number of positions (instead of zero), and removing zero as a
default argument for num_positions. Nobody should call this method
with zero positions or velocities, and those certainly shouldn't be
our default values.