multibody/tree: specify any axis for ScrewJoint
Adds a constructor to ScrewJoint that accepts a
Vector3<double> for the axis unit vector and a
screw_axis() accessor method to the class. The
current constructor that assumes the axis unit
vector is the Z-axis is retained but may be
considered for deprecation in the future. The
constructor implementation is moved to the .cc
file and throws a std::logic_error if invalid
parameters are supplied.

Adds a Vector3<double> argument to ScrewMobilizer
that is required to have a non-zero norm. The
mobilizer math is updated to use this axis instead
of assuming the Z-axis.

screw_mobilizer_test: a bug in the
CalcAcrossMobilizerSpatialVeloctiy method and test
is fixed to use angular velocity as generalized velocity
It now matches the way spatial acceleration is
implemented and tested. An additional expectation on
the product of force and velocity computed in
different frames is added.