Improve/fix obb hierarchy construction (#17937)
We fit an OBB to a mesh (or cluster of features) using PCA analysis. It
provides a reasonable estimate of a good OBB orientation. We have an
optimization step that attempts to create a *better* orientation (and,
thereby, a better fit). To optimize, we compute the partial derivative of
OBB volume w.r.t. orientation of the box.

Previously, it was possible for that partial to have *zero* magnitude. This
led to failures because the optimization algorithm would end up with
infinite values (which could, in turn, lead to NaN values). So, we have to
protect ourselves against that.

But beyond that, if the magnitude of the partial derivative is sufficiently
small, there's really very little value in trying to further optimize the
fit. So, we apply a threshold test for optimization to skip the effort if
it doesn't seem worth it, and protect ourselves from introducing bad
inf and NaN values.

Incidentally, also moved the constructor to the .cc file where it belonged
all along.