[IRIS] New termination conditions for IrisInConfigurationSpace
Adds options.num_collision_infeasible_samples to support a
probabilistic guarantee of correctness (with probability one as the
number of samples goes to infinity) for the
nonlinear-optimization-based version of IRIS.

Because this requires solving potentially many more (nearly identical)
MathematicalPrograms, I've finally done the refactor which reuses the
allocated programs and just updated the constraints.

BREAKING CHANGE: We have elected to remove support for Ibex from IRIS;
it proved to only be useful for exceptionally toy problems and has
become a footgun for users that turn it on (and wait forever). Rather
than carry it through this refactor, I have elected to remove it.