Improve error reporting when pydrake import fails
Modify pydrake (at least, `import pydrake.all`) to detect when it can't
be loaded due to failure to load a dynamic library, and to give a
somewhat more user-friendly error message. Add a brief document listing
what distro packages need to be installed, which is referenced by the
aforementioned error.