Support end-effector velocity limits in DiffIK and integrate through kStuck
Adds end-effector velocity limits for the entry point which takes a
desired pose.

DiffIKIntegrator now continues to integrate even if DiffIK reports
kStuck. Previously, kStuck (which means small velocity commands)
resulted in the integrator commanding exactly zero velocity. This
meant it was effectively impossible to get unstuck using DiffIK. The
only reason for this before was because DiffIK did not return joint
velocities when it was kStuck; but that limitaiton was resolved with
my last PR.

Also adds a few minor improvements to the differential inverse
kinematics tools:
- Minor touch-ups / todos in the documentation.
- Use the desired pose entry point in DiffIKIntegrator.
- Strengthens the TestLinearObjective test.

These changes will not impact the behavior for current users of
DoDifferentialInverseKinematics; they only add new options.