[doc] Fix description list handling for pydrake (#18006)
Handle dl/dt/dd tags in the pydrake doc toolchain by translating them
into bold titles and stripping the rest. This fixes a visual bug in the
pydrake.visualization docs.

Also fix a grammar error that was irking me and clarify that --quick
in the doc build instructions only works for Doxygen.