[workspace] Update qdldl to latest release 0.1.6
Even though the README advises not to upgrade qdldl past osqp's pin of
qdldl (currently near 0.1.5), the changes from 0.1.5 to 0.1.6 are only
to the build system so we might as well take them now to make any future
feature-rich upgrades easier.

Adjust github paths to match relocated qdldl and osqp repositories.

Opt-in to strict mode for cmake configure variables.