Move Joint::type_name() implementations to their corresponding source files.
Because this function body is in a header file and uses static never_destroyed,
it in effect creates global storage that could end up having definitions in
multiple object files. Having consulted the local C++ experts, we believe that
a pure C++ environment could handle this hazard, but that it could potentially
cause subtle multiple-definition problems for Drake's python bindings shared
libraries. That is, individual libraries might build, but have symbol
conflicts when loaded together in a single program.