[pydrake] Revert "Don't load native code upon 'import pydrake'"
This reverts commits 9ff5cc1d and 6d80c692, though a few incidental
improvements are kept intact (not reverted).

Since the time that code was written, our plans for how to deal with
Drake's Python lcmtypes have evolved. The attempt to move lcmtypes into
the same module as Drake itself is now much further down the line, and
will be preceded by a much larger reorganization of the pydrake native
code first (folding our modules down into a single shared library).

Maintaining the "no native code" invariant in the meantime is too
difficult. In particular, we'd like to give better error messages when
loading our native code fails, and having that in one place will make
that job much easier. We can resurrect this test (and invariant) when
we circle back to changing the lcmtypes' packaging.