Adds `drake_gym` to Drake for OpenAI Gym compatibility
This commit brings Drake Gym into Drake.  This is a tool for
turning a Drake Diagram into an OpenAI Gym Env.  Two examples
are provided to clarify this rather complex operation.

Note:  This commit includes `gym` and a patched version of
`stable_baselines3` to allow tests to run.  However this is not
a full version of `stable_baselines3`; users who want to use that
will need to install it themselves.

Co-Authored-By: Jose Barreiros <jose.barreiros@tri.global>
Co-Authored-By: Russ Tedrake <russt@mit.edu>
Co-Authored-By: Jeremy Nimmer <jeremy.nimmer@tri.global>