[pydrake] Bind dataclass-like schemas for serializable structs (#18109)
C++ structs whose Python attributes are bound using Serialize now offer
a __fields__ static property, enabling other pydrake code to interrogate
the static schema of the type (akin to Python's dataclasses fields).

One consequence of this change is that the C++ classes used as struct
fields must be bound prior to binding the struct that uses them. The
CameraConfig structs need an adjustment to accommodate this.