Alphabetize shapes
The ordering of the shapes has historically been chaotic (they were ordered
according to introduction order). This makes it hard to find the desired
shape and interferes with assessing if a shape implementation is missing
in a reifier.

The declarations of shapes are now alphabetized as well as the in the
ShapeReifier implementation. Tests and implementations have been updated to
reflect the same order.