[meshcat] Add support for gamepads in Meshcat
The meshcat javascript reads the gamepad info and sends it to the
meshcat server whenever a change is detected.  The c++ meshcat
interface provides access to the most recent gamepad status.

This feature is provided primarily to support applications where Drake
is running on a remote machine (e.g. in the cloud), and the Meshcat
javascript in the browser is the only code running on the local
machine which has access to the gamepad.