Fix a missing paramname
This also moves the default annotation for this argument to allow the
subsequent return annotation to create proper output, at the cost of
mangling the pydoc output of this default directive. But at least this
is semi-readable and this exact problem exists with default annotations
elsewhere in our pydocs and I'm looking into it.