[controllers] InverseDynamics doxygen and scalar support
Cleans up the doxygen for InverseDynamics and
InverseDynamicsController.

Adds support for default scalars to both classes, and implements the
scalar conversion. (This required moving the owned plant from
InverseDynamicsController to InverseDynamics, which is where it really
belonged).

Also fixes up a bunch of errant uses of @tparam_default_scalar throughout
the codebase that I found while trying to remember how to spell it!