[pydrake] Switch the clang-format directives order
The config file is "last one wins" so pydrake's dotfile should be
appending to the root settings, not prepending.

This hasn't mattered in the past because we had not yet overridden any
settings in pydrake that were also set in the root; that might be
changing soon, so we'll preemptively correct the order here now.