[meshcat] Upgrade Meshcat and fix Meshcat::StaticHtml()
My recent meshcat changes broke the StaticHtml()
method. meshcat_manual_test revealed the problem but I failed to
notice during my offending PR! I've fixed it upstream and this PR
updates the meshcat SHA.

Also updates the gamepad logic to not throw an error due to not having
a connection.

Resolves #18342.