[parsing] Initial collision filter support in mujoco parser
Adds support for parsing the `<contact>` elements.

Adds an accessor method to MultibodyPlant to get the registered
SceneGraph.  I don't know why we've never offered it before; I don't
see a problem with it.  And I can't get access to the collision filter
logic directly without it.  The CollisionFilterResolver hoopla from
the other parsers isn't immediately suited to this MuJoCo case.