[wip] test_gltf_conversion does not appear to be valid
After skipping some broken POSITION vs TEXCOORD texture unit matches
and `indices` as well, we end up with invalid matrix comparisons
as well as numerous different sized buffer objects etc.

It is not clear how to get the original data out for comparison
to inspect manually, this test being broken on macOS may indicate
that the client is broken on macOS entirely?