GurobiSolver returns the solution from Gurobi at termination.
This solution is the primal and dual variable values when Gurobi
terminates. The solution is reported when Gurobi solver doesn't
encounter an error, including optimality/infeasible/unbounded/hitting iteration
limit, etc. Note that if Gurobi encounters an error, we set the
solution result as InvalidInput and don't set the
MathematicalProgramResult solution.