Allow pressing Escape to stop ModelVisualizer
Also change the message that ModelVisualizer prints so that it is
accurate when running in a notebook, under the theory that someone
running the script from the command line will understand that they can
also press Ctrl-C to quit. This brings this script more into line with
JointSliders::Run.

Fixes #18348