[meshcat] Hoist recording logic from MeshcatVisualizer to Meshcat
Towards #17832 -- with this PR we now support recording contact
visualization in meshcat animations (only not the hydroelastic surface
visualization yet).  Previously, contact forces did not appear in any
animations.

Previously the recording logic lived in the System
(MeshcatVisualizer), because it was the system that knows about
time. But this became unwieldy... as more systems wanted to contribute
to the animation (for instance the contact visualizer, the point cloud
visualizer, my manipulation class meshcat writer, ...), I had to start
writing a significant amount of custom recording-specific logic in
each subsystem. (I do have a working experimental branch in which I
made this possible through a recording output port on the
MeshcatVisualizer... but the burden on every downstream visualizer was
too great).

In this PR I've hoisted the recording logic from MeshcatVisualizer
into the main Meshcat class.  SetTransform and SetProperty can take an
additional optional input for time_in_recording. If a user has called
StartRecording, then any future calls to SetTransform/SetProperty with
the time included will be recorded.