[primitives] Add TrajectorySource::UpdateTrajectory
I've seen a number of examples where a user wants to construct a
Diagram in order to compute quantities for planning, then execute the
planned trajectory using a TrajectorySource. This caused a
chicken-and-the-egg problem -- they couldn't produce the final diagram
until they had the trajectory but couldn't produce the trajectory
until they had the Diagram. Constructing a completely different
Diagram for the planning felt too cumbersome.

With the addition in this PR, a user can set up the Diagram with a
placeholder trajectory that only needs to be the right size, then set
the trajectory after the Diagram is built.