SceneGraph allows changing geometry shape
In addition to changing things like collision filters or geometry
properties after registration, now the Shape associated with a geometry
instance can also be changed.

This is *especially* useful if a user wants to change a body in an existing
plant. They can change the mass properties via MbP's mass "parameters" and
change the geometric representation in SceneGraph to match.

The new shape is associated with the original GeometryId, making it friendly
to MbP's internals.

This required some refactoring of the GeometryState code that controls the
registration/removal of geometry with respect to the proximity and render
engines.