Deprecate "register w.r.t. geometry" SceneGraph APIs
Users have the ability to register a geometry in a fixed pose relative to
another geometry. Ultimately, it has a fixed pose relative the common
geometry::Frame. This extra layer has not proved to be useful and the
concept of "parent vs frame" has only caused confusion.

This deprecates the limited public API in preparation for removing all of
the supporting internal code.