[visualization] Add sliders for geometry alpha in MeshcatVisualizer
Adds optional sliders to adjust the visual alpha of geometry added by
MeshcatVisualizer. These sliders default to being disabled for now while
we discuss the preferred method for handling these without breaking
other code.

This reworks #18454, which was reverted by #18463; the main difference
is that the sliders are disabled by default, and so the dubious test
deltas aren't necessary.