Add VTK textured capsule source.
Replace the vtkCapsuleSource with vtkTexturedCapsuleSource. This
implementation contains the texture coordinates. It also changes the wireframe
construction of the capsule. Now a single sphere is used where the poles are
parallel to the cylinder body. The vtkCapsuleSource has two spheres with their
poles parallel to the normal of the cylinder body. This change enables
adding texture to the capsule body.