[C-IRIS] Consolidate to the formulation using Chebyshev center for polytope. (#18496)
If I were to use the formulation that doesn't introduce Chebyshev, but a'*x + b >= 1, then this causes problem in the bilinear iteration. Since when we search for the separating margin, we use the formulation a'*x + b - margin - lambda(s) * (d - C*s) is sos. Then in the next iteration when we search for C and d, we don't need the margin and use the formulation a'*x + b - 1 - lambda(s) * (d-C*s) is sos. This causes the problem that the solution a, b, lambda in the first iteration with margin is NOT a feasible solution for the next iteration when we search for C and d but fix lambda(s) (as very likely the margin is < 1).