[setup] Avoid accidentally installing OpenBLAS
Drake by default should stick with the Ubuntu default BLAS (Netlib).
Users are invited to install a custom BLAS, but we shouldn't force it.

Through a chain of calamities in upstream Ubuntu 22.04 packaging,
Drake's binary_distribution dependency on python3-matplotlib leads to
the dependency chain of:
  python3-matplotlib =>
  python3-fonttools =>
  python3-scipy | python3-munkres
with
  python3-scipy =>
  python3-pythran =>
  libopenblas-dev | ... | libblas.so

This ends up installing the OpenBLAS dev package, trumping our
source_distribution dependency preference for libblas-dev that
comes along in the second install command.

Work around this problem by elevating Drake's BLAS dev package
dependency from a source dep to a binary dep, which is able to
guide python3-pythran's disjunction to use Netlib.

Furthermore, now that we see scipy in the chain (and knowing that
Drake goes to great lengths to try to require scipy), add a binary
dependency on python3-munkres to short-circuit that disjunction in
order to reduce our footprint.

Technically munkres on its own (without the libblas-dev change) would
be sufficient to get us to Netlib BLAS, but we'll keep the dev package
in binary deps to preempt similar potholes in the future.