Work around build issue
Something in the build is wrongly "de-duplicating" link arguments, possibly only on the scope of rules for a particular library(?), which is transmuting e.g. `[-Wl,]-framework [-Wl,]A [-Wl,]-framework [-Wl,]B` into `[-Wl,]-framework [-Wl,]A [-Wl,]B`, which makes the linker unhappy because `B` without `-framework` is (not surprisingly) broken. Work around this by combining these into the form `-Wl,-framework,X`, which remains unique until actually passed to the linker.

For some reason, this is newly a problem with this PR, and only seems to affect wheel builds.