[pydrake] Major rework of template classes
When a template class has a default type parameter, name the default
instantiation directly using the default name, instead of via an alias.
This makes IDE auto-complete and type annotations more natural.

For template instantations, use Unicode mangling instead of square
brackets for the underlying class name. Python class names are not
allowed to use brackets; doing so cases headaches for tools like Mypy.
On the other hand, for the website we always want the demangled names
so turn off name mangling during docs generation via a giant hack.

When displaying possibly-mangled class names (e.g., in repr), as much
as practical in the bindings use the demangling function to provide
the more natural expression for users to see.

Fix RollPitchYaw repr for non-floats vs a missing template argument.