[pydrake] Introduce name mangling for template classes
For template instantations, use Unicode mangling instead of square
brackets for the underlying class name. Python class names are not
allowed to use brackets; doing so cases headaches for tools like Mypy.
On the other hand, for the website we always want the demangled names
so turn off name mangling during docs generation via a giant hack.

When displaying possibly-mangled class names (e.g., in repr), as much
as practical in the bindings use the demangling function to provide
the more natural expression for users to see.