[pydrake] Introduce name mangling for template classes
For template instantations, use Unicode mangling instead of square
brackets for the underlying class name, and similarly for commas and
periods within the template param. Python class names are not allowed
to use brackets; doing so causes headaches for tools like Mypy.  On
the other hand, for the website we always want the demangled names so
turn off name mangling during docs generation via a giant hack.

When displaying possibly-mangled class names (e.g., in repr), as much
as practical in the bindings use the demangling function to provide
the more natural expression for users to see.

Add a regression test that Mypy can parse our *.pyi files (modulo one
file that still needs some grooming).