[tools] Upgrade benchmark_tool to Jammy (#18631)
Bump the supported Ubuntu version within benchmark_tool to Jammy 22.04.

Also, make the routine for checking installed package versions a bit
more robust. If the package of interest is in a broken state, the
dpkg-query output may not contain a version number at all. The revised
code avoids raising an exception in that case.