[pydrake] Fix incompatibility with Mypy
When parsing the docstrings of a function to find its set of overloads,
C++ code samples that mention the function name were at risk of matching
the C++ scoping operator "::" as a type annotation, leading to malformed
output. For example "MyFunc(foo: int) -> bool" is a valid signature, but
the sample code "MyFunc(Eigen::Lower)" was being mis-parsed. Fix this by
patching stubgen to reset in case it sees a double colon.

Enable the stubgen regression for all of pydrake; no more caveats.

Revert da58ae08 and 2cf04fdc header file changes, which were work-arounds
for this parsing bug and are no longer needed.