[multibody] Add AddModelsFromUrl function to Parser
Overload in pydrake as AddModels(url=...) sugar.
Port all examples and tutorials to use this spelling.

Users are constantly confused by the drake-specific FindResourceOrThrow
in our code when their go to write their own software with non-Drake
model files. For parsing, the better answer is always the PackageMap,
so we should make it easy to load models via package:// URLs.