[planning] Simplify RobotDiagramBuilder function names (#18772)
The typical use case of a builder is that the object is mutable and
users are mutating it. In that case, the clarity benefit of using
"mutable_" prefix on is small compared to the cost. (In contrast with,
e.g., a Context where mutation access causes cache entry invalidation
so using special names make sense.)

Also remove the IsPlantFinalized and FinalizePlant sugar. The sugar
isn't encapsulating or abbreviating anything versus calling the MbP
functions directly on the plant() accessor.