ModelVisualizer: re-pose joints by name
Makes the reload feature of ModelVisualizer more robust by setting joint
positions in the reloaded model to the values of joints with the same
names, rather than requiring an exact match of the number of joints
before and after.

Also fixes docstrings relating to accessing the visualizer's parser, and
warns the user if the reload option is invoked when the parser was
accessed directly. Also adds a package map accessor method so that the
main script doesn't need to access the parser method in order to get to
the package map.