ModelVisualizer: re-pose joints by name
Makes the reload feature of ModelVisualizer more robust by setting joint
positions in the reloaded model to the prior values of joints with the
same names, rather than requiring an exact match of the number of joints
before and after.

Adds a new method to Meshcat to fetch all slider names to enable the
above behavior.

Also fixes docstrings relating to accessing the visualizer's parser.