[parsing] Integrate ParsingWorkspace into Mujoco parser
This patch clears the way for making the Mujoco parser work more like
the other parsers in the suite. It will be necessary for integrating
control of diagnostic output, and for other forthcoming changes that
affect all parsers.

The bulk of the changes are in the unit test, but that should all be
functionality-neutral.