Improve defaults for AffineSystem and LinearSystem
It's always been a pain to produce simple linear/affine systems,
because we needed to pass in empty matrices of the correct size
(e.g. even if we have no states, the B matrix must have zero rows and
number of input columns).  This PR makes the constructor for
AffineSystem more robust, setting any empty matrix to an appropriat
size. This makes it reasonable to have default values.

This motivation for this was that I was writing a tutorial where the
simplest way to produce a system that simple adds an offset to a
signal (e.g. to change coordinates), took many lines and lots of
explaining. Now we can just have e.g.

```
system = AffineSystem(D=np.eye(2), y0=[1, 2])
```