Correct and unify examples in pydrake docstrings
We have several pure python methods with example code in their doc
strings. In some cases, they don't render as doc strings at all. In other
cases, they render inconsistently (i.e., as a list).

This picks an arbitrary style for the examples and confirms that the
instances all render the same.