[tools] Keep bazel analysis phase intact (#18838)
As it turns out, varying 'test' vs 'build' for options in a bazelrc
causes the analysis phase to be re-run as the user alternates between
'bazel build' and 'bazel test'.  Since all test options are valid for
build as well, we should just list everything in build.

This removes "Options have changed, discarding analysis cache" output
from bazel in the typical case.