[symbolic] Speed up Add and Mul cell ctors
Note that expression_cell.h is internal use only, so this API change
will not disturb any users.

This is roughly a 50% reduction in BenchmarkMatrixInnerProduct/200,
and a 10% reduction in all three BenchmarkSosProgram.

Also move two tiny Variable getters to the header file. It's a little
difficult to show this in a profile, but I think we'd all agree that
this change is blindingly obvious and should have been like this from
the beginning.