[solvers] Change SolverBase constructor to prefer SolverId directly
The constructor with a SolverId-returning functor is now deprecated.

Fetching the solver_id() should be a zero-cost abstraction, which
means the function must be final and inline, rather than a virtual
function that makes another indirect call.

Add an API comment and runtime warning about SolverId pessimizations.
The SolverId is supposed to be a integer-like flyweight that's cheap to
pass by value, but when its name exceeds 15 characters (i.e., more than
SSO can handle) then we were using heap when copying it (e.g., into the
SolverOptions maps).

While we're updating comments anyway, upgrade them all to the preferred
lightweight Doxygen style.