Eliminate unnecessary heap allocation in Geometry_properties
The private method GetValueOrThrow would take the name of the calling
function in the event of failure so it could be included in the error
message. However, that method name was being captured as a
const std::string&. Going from string literal to string requires a copy
of the string (and a concomitant heap allocation).

By replacing string with string_view, we eliminate the copy.