[symbolic] Speed up symbolic::Variable a little bit
Pack the Variable::Type into the ID. This reduces sizeof(Variable) from
32 to 24 bytes.

This is about a 2-3% speedup on BenchmarkPolynomialEvaluatePartial, but
the main benefit is a step towards a size of 16 bytes with SSO in a
future commit.

This changes the order of variables in ordered collections. Previously,
they were sorted by creation time. Now they're effectively sorted by the
pair of (variable type, creation time).