Improve new_release script.
Fix new_release so it can be run via `bazel run`. Teach it to recognize
notices that should be displayed during an upgrade. Add markers to
externals that have such information. Update documentation.

This should avoid the need for the user running the script to manually
inspect each `repository.bzl` to determine if additional actions need to
be performed.