Improve new_release script.
Fix new_release so it can be run via `bazel run`. Add a field to
github_archive to communicate advice to new_release that should be shown
to the user when performing an upgrade. Use this instead of comments in
relevant externals. Update documentation.

This should avoid the need for the user running the script to manually
inspect each repository.bzl to determine if additional actions need to
be performed.