Modernize emacs bazel-mode hints
For *.bzl and *.bazel (with no dots in the *), the mode is
auto-detected; remove the hint.

For package.BUILD.bazel and template files with odd extensions, switch
the hint from python-mode to bazel-mode.

Remove boilerplate comments for what a BUILD file is; they made sense
five years ago but are superfluous now.

Also remove the analogous vi python-mode hints.