[perception] PointCloud can morph its fields after construction (#18987)
Prior to this PR, the `Fields` of a PointCloud object is determined
at construction time, and doesn't allow any morphing after that.
An exception will be thrown if one attempts to copy/move-assign a
PointCloud with a different `Fields` to another. However, the same
PointCloud can be used to copy/move-construct another Pointcloud.
This PR tries to make the behavior consistent and avoid confusions
for users.