[fem] Remove multi-thread test for PETSc (#19073)
The upgrade from Clang-12 to Clang-14 reveals that more atomic patching
is needed for PETSc for thread safety. This has become tedious as PETSc
is not designed to be thread safe to begin with.

As we have plans to remove the PETSc dependency in the near future, we
remove the failing multi-thread test for now instead trying to apply
more band-aid to stop the hemorrhage.