[multibody] Add Block3x3SparseMatrix (#19000)
Block3x3SparseMatrix is a highly-specialized sparse matrix data
structure that stores data as 3x3 dense blocks. It provides a minimal
set of APIs to support the matrix operations required for constraint
Jacobians in SAP. It is particularly useful for constraints involving
deformable bodies where the non-zero entries usually assume a 3x3 block
structure (i.e. each constraint has 3 constraint equations and dofs
involved in the constraint also come in groups of 3 as the x,y,z
coordinates of the deformable mesh vertex involved in the constraint).