[parsing] Don't abort on invalid input (#19125)
Add missing IsValid check to the AddModelInstance directive.

DRAKE_DEMAND is an inappropriate tool to reject malformed input. We
should prefer exceptions rather than taking down the whole Python
interpreter.

Now that GetModelInstanceByName has a nice error message, we should
rely on that rather than an ahead-of-time check in the parser.