[tutorials] Cite new tools for OBJ loading
Add text and examples of Parser.AddModels on an OBJ file.

Shorten simulation duration to run in <1 minute in Debug builds. Also
record 2 seconds (not 5) for live users; nothing happens after that.

Add link to mesh_to_model and obj2mjcf.

Update the rendering tutorial to use the colorizer systems.