[parsing] Port mujoco parser to diagnostic policy (#19157)
This patch updates most of the mujoco parser diagnostics to include file
and line number, and to obey policy configuration from the top-level
Parser.

In addition, it fixes a bug found trying to use std::optional<bool> for
a tri-state variable -- one dropped '*' in the wrong place led to an
unreachable branch, and probably other problems.

Finally, it moves the affected files closer to clang-format compliance.