QueryObject allows signed distance between a filtered geometry pair
Historically, requesting signed distance between two geometries that had
been filtered would throw. While it might make sense to have the geometries
filtered for contact, it might still be convenient to know how far they
are. This changes the API to allow the caller to demand that collision
filters be ignored.