[visualization] Use a mutable scene_graph for setup
Forthcoming visualization code will need to register new illustration
geometry, and passing the scene_graph by const pointer is misleading
in this case. (Even though we already have a mutable SceneGraph in the
DiagramBuilder, the const here hinted that we would not mutate it.)

Deprecate the const scene_graph overload of ApplyVisualizationConfig.