[parsing] Improve SDFormat inertia invalidity reporting
This patch brings inertia invalidity reporting under diagnostic policy,
and relaxes inertia problems from hard throws to warnings.

The move to diagnostic policy means that inertia warnings give file
names and line numbers, which make it easier to fix problems. In the
case of SDFormat, this patch demotes the useless error message from the
underlying sdformat library to a warning, and emits more detailed
warnings in the later translation-to-Drake phase.

The move from hard throws to warnings means that models with
non-physical inertias can still be used with non-simulation workflows,
such as visualization, kinematics, and analysis.