[meshcat] Enable alpha adjustment of invisible geometry
If geometry is invisible, the meshcat alpha sliders will never make it
visible because MeshcatVisualizer simply multiplies geometry alpha by
the slider value. To improve this somewhat, check if all geometry in a
visualizer is invisible and if so use the absolute alpha slider value as
the new alpha instead.

This commit also changes some of the indent whitespace for
meshcat_visualizer_params.h to be consistent with itself and the sibling
headers.