Fix ConvexSet move operations
Add now-required virtual DoClone to the base class; deprecate the
ConvexSetCloner function and the ConvexSet constructor that accepted it.

Fix some DEMAND checks on user input to be THROW instead.
Invalid arguments should result in exceptions, not aborts.

Fix documentation and implementations to be precise and correct in the
case of zero-dimensional sets.