Filter contact results for completely locked trees.
This results in a breaking change: the members of `ContactResults` will
no longer be in bijection with the geometry quantities
(`PenetrationAsPointPair` and `ContactSurface`) that inform the solver
of contact. Geometric contact data between completely locked trees will
simply be ignored. Any program that made the assumption that:
  num_contact_results = num_point_pairs + num_contact_surfaces
will be broken.