ShapeReifier can configure default behavior
Historically, the default behavior for an unimplemented Shape in a
ShapeReifier has been to throw. It was possible for derived reifiers to
tweak the message thrown, but the behavior was fixed.

This is very clunky of a reifier wants to make use any API that takes a
const Shape& as its type. The call would have to be explicitly implemented
across all types.

This adds the DefaultImplementGeometry() virtual method. Its implementation
is the same historical behavior: throw with whatever message is provided.
But now it can be overridden so that the reifier can redirect an arbitrary
subset of shape types to another API as default behavior.