[workspace] Build IPOPT from source on Ubuntu
On Ubuntu, this is an upgrade from 3.11.9 to 3.14.12.

On Ubuntu, we can now remove ipopt from the prereqs setup and skip it during
wheel builds. It's licensed as EPL-2.0 so we can statically link it but we do
need to distribute a patch file with our changes. We also teach vendor_cxx
about gentler option for vendoring, where we only hide the symbols without
prefixing them with "drake_vendor".

On macOS, Homebrew usually tracks the newest upstream version so the urgency of
rebuilding from source is reduced. Building from source would still help with
linker issues, but since there is no MUMPS package in Homebrew the work to make
it happen is not yet worth the time investment.