Improve math::AreAutoDiffVecXdEqual
Previously, it was possible that ExtractGradient (and therefore
AreAutoDiffVecXdEqual) would throw, when this method should simply
return false. Now we handle the case where the elements of the vector
could have a different number of derivative values.

This version should also be slightly faster.

N.B. This came up for me because I had two different constraints
setting values in a shared Context, but one was calling
SetPositions(), and the other was calling SetPositionsAndVelocities().